@@ -1,5 +1,5 @@
 module nibbled
 
-go 1.13
+go 1.16
 
 require github.com/armon/go-radix v1.0.0
