@@ -3,21 +3,22 @@ package main
 import (
 	"bufio"
 	"bytes"
+	"fmt"
 	"io"
 	"os"
 	"path/filepath"
 )
 
-func loadGopherMap(out io.Writer, localPath string, selector string) error {
+func loadGopherMap(out io.Writer, localPath, selector string) error {
 	f, err := os.Open(filepath.Join(localPath, "gophermap"))
 	if err != nil {
-		return err
+		return fmt.Errorf("could not open gophermap: %w", err)
 	}
 	defer f.Close()
 	return processGopherMap(f, out, localPath, selector)
 }
 
-func processGopherMap(in io.Reader, out io.Writer, localPath string, selector string) error {
+func processGopherMap(in io.Reader, out io.Writer, localPath, selector string) error {
 	reader := bufio.NewReader(in)
 	for {
 		line, _, err := reader.ReadLine()
@@ -25,28 +26,35 @@ func processGopherMap(in io.Reader, out io.Writer, localPath string, selector st
 			break
 		}
 		if err != nil {
-			return err
+			return fmt.Errorf("cannot read from gophermap: %w", err)
 		}
 		if bytes.HasPrefix(line, []byte("#")) || bytes.HasPrefix(line, []byte("!")) {
 			continue
 		}
 		if bytes.HasPrefix(line, []byte(".")) {
-			out.Write([]byte(".\r\n"))
+			if _, err := out.Write([]byte(".\r\n")); err != nil {
+				return err //nolint:wrapcheck
+			}
 			break
 		}
 		if bytes.HasPrefix(line, []byte("*")) {
-			if catalogue, err := listDirectory(localPath, selector); err != nil {
-				return err
-			} else {
-				_, err := write(out, catalogue)
-				return err
+			catalogue, err := listDirectory(localPath, selector)
+			if err == nil {
+				_, err = write(out, catalogue)
 			}
+			return fmt.Errorf("could not list directory: %w", err)
 		}
 		if !bytes.ContainsRune(line, '\t') {
-			out.Write([]byte("i"))
+			if _, err := out.Write([]byte("i")); err != nil {
+				return err //nolint:wrapcheck
+			}
 		}
-		out.Write(line)
-		out.Write([]byte("\r\n"))
+		if _, err := out.Write(line); err != nil {
+			return err //nolint:wrapcheck
+		}
+		if _, err := out.Write([]byte("\r\n")); err != nil {
+			return err //nolint:wrapcheck
+		}
 	}
 	return nil
 }
